+++ src/main/java/org/example/model/dto/response/DepositResponseDTO.java
@@ -9,12 +9,12 @@ import java.time.LocalDateTime;
 
 @Data
 public class DepositResponseDTO {
-  private Long moneyAmount;
+  private int moneyAmount;
   private CurrencyType currencyType;
   private DepositStatus depositStatus;
   private LocalDateTime creationDate;
   private DepositType depositType;
-  private Float depositLimit;
+  private int depositLimit;
   private LocalDateTime incomeDate;
   private LocalDateTime expireDate;
 }
